// types.ts
export enum Tab {
  Dream = 'Dream Visualizer',
  Identity = 'Identity Styler',
  Icons = 'Meet Your Icons',
  Art = 'Art Styler',
  Fusion = 'Image Fusion',
  Commercial = 'Commercial Creator',
}

export enum DreamStyle {
    Manga = 'Manga',
    Cinematic = 'Cinematic',
    Storybook = 'Storybook',
    Surreal = 'Surreal',
    Fantasy = 'Fantasy',
}

export enum ArtStyle {
    Ghibli = 'Ghibli Studio',
    Impressionism = 'Impressionism',
    Cubism = 'Cubism',
    PopArt = 'Pop Art',
    Steampunk = 'Steampunk',
    Synthwave = 'Synthwave',
    Custom = 'Custom...',
}

export enum CommercialStyle {
    Photorealistic = 'Photorealistic',
    Vintage = 'Vintage Ad Style',
    Animated = '3D Animated (Pixar Style)',
    Minimalist = 'Modern & Minimalist',
    Futuristic = 'Sleek & Futuristic',
    Watercolor = 'Artistic Watercolor',
    Custom = 'Custom...',
}

export type Persona = string;
export type Icon = string;